+++ src/shared/environment/EnvironmentLocator.php
@@ -9,6 +9,9 @@ class EnvironmentLocator {
      * @return Environment
      */
     public function getEnvironment($operatingSystem) {
+        if (strtolower($operatingSystem) === 'darwin') {
+            return MacOsEnvironment::fromSuperGlobals();
+        }
         if (strtoupper(substr($operatingSystem, 0, 3)) === 'WIN') {
             return WindowsEnvironment::fromSuperGlobals();
         }