+++ src/shared/exceptions/FileNotWritableException.php
@@ -1,0 +1,6 @@
+<?php
+namespace PharIo\Phive;
+
+class FileNotWritableException extends \Exception implements Exception {
+
+}