+++ src/shared/exceptions/LinkCreationFailedException.php
@@ -1,0 +1,6 @@
+<?php
+namespace PharIo\Phive;
+
+class LinkCreationFailedException extends \Exception implements Exception {
+
+}