+++ src/shared/file/SymlinkPharActivator.php
@@ -12,8 +12,21 @@ class SymlinkPharActivator implements PharActivator {
      * @return Filename
      */
     public function activate(Filename $pharLocation, Filename $linkDestination) {
+        $this->ensureDestinationIsWritable($linkDestination);
+
         symlink($pharLocation->asString(), $linkDestination->asString());
         return $linkDestination;
+    }
+
+    /**
+     * @param Filename $destination
+     *
+     * @throws FileNotWritableException
+     */
+    private function ensureDestinationIsWritable(Filename $destination) {
+        if (!$destination->getDirectory()->isWritable()) {
+            throw new FileNotWritableException(sprintf('File %s is not writable.', $destination->asString()));
+        }
     }
 
 }